Add files via upload
The modelling and prediction of petroleum prices is an important factor for economic growth in developing countries such as Kenya. It is crucial to understand the current and future prices of petroleum as it can help users make informed decisions about buying and selling patterns. Additionally, it assists the government in implementing appropriate policies to maintain stable prices. The study utilized the Auto-regressive integrated moving average model (ARIMA) to account for the volatility in petroleum prices and predict future prices. The study discovered that the monthly petroleum prices in Kenya were non-stationary, indicating instability within the petroleum market. The ARIMA (3,2,2) model was deemed suitable and valid for predicting volatility and forecasting petroleum prices. To enhance economic growth and assist users in making informed decisions, the study recommends that the Kenyan government should implement a stable and low petroleum pricing system. This would minimize production costs as petroleum is one of the inputs in production. The ARIMA model was then used to predict petroleum prices for the period between January 2011 and December 2021. Which was necessary in forecasting the future prices in coming years